Restructure packages.
